map view updates on step list scroll
