+++ Footprints/Trips/StepView.swift
@@ -9,10 +9,10 @@ import SwiftData
 import SwiftUI
 
 struct StepView: View {
-    @Environment(\.modelContext) private var modelContext
+    
     @Bindable var trip: Trip
-    @State private var position: PersistentIdentifier?
-    @State private var currentStep: Step?
+    @Binding var position: PersistentIdentifier?
+    
 
     var body: some View {
 
@@ -42,21 +42,11 @@ struct StepView: View {
         .navigationDestination(for: Step.self) { step in
             Text(step.timestamp, style: .date)
         }
-        .onChange(of: position) {
-            let step = getStep(for: position)
-            print(step?.timestamp.formatted(date: .abbreviated, time: .shortened) ?? Date.now)
-        }
+
     }
 
-    func getStep(for id: PersistentIdentifier?) -> Step? {
-        guard let id else { return nil }
-        guard let step = modelContext.model(for: id) as? Step else {
-            return nil
-        }
-        return step
-    }
 }
 
 #Preview(traits: .previewData) {
-    StepView(trip: .bedminsterToBeijing)
+    StepView(trip: .bedminsterToBeijing, position: .constant(Step.bedminsterStation.persistentModelID))
 }
